requirements.txt
# Core dependencies
streamlit>=1.30.0
python-dotenv>=1.0.0

# LangChain and related packages
langchain>=0.1.0
langchain-community>=0.0.12
langchain-groq>=0.0.4
langchain-huggingface>=0.0.2
langchain-text-splitters>=0.0.1
langchain-ollama>=0.0.1

# Document processing
pypdf>=3.17.1
pdfplumber>=0.10.3
unstructured>=0.11.0
markdown>=3.5.1

# Embeddings and vector stores
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Data processing
pandas>=2.1.3
numpy>=1.26.2
matplotlib>=3.8.2

# Utilities
tqdm>=4.66.1
requests>=2.31.0
watchdog>=3.0.0

# Testing
pytest>=7.4.3